third_party/WebKit/LayoutTests/external/wpt/selection/removeAllRanges-expected.txt

CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
CONSOLE ERROR: line 944: The given range isn't in document.
This is a testharness.js-based test.
Found 118 tests; 88 PASS, 30 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Range 0 [] forwards 
PASS Range 0 [] backwards 
PASS Range 1 [paras[0].firstChild, 0, paras[0].firstChild, 0] forwards 
PASS Range 1 [paras[0].firstChild, 0, paras[0].firstChild, 0] backwards 
PASS Range 2 [paras[0].firstChild, 0, paras[0].firstChild, 1] forwards 
PASS Range 2 [paras[0].firstChild, 0, paras[0].firstChild, 1] backwards 
PASS Range 3 [paras[0].firstChild, 2, paras[0].firstChild, 8] forwards 
PASS Range 3 [paras[0].firstChild, 2, paras[0].firstChild, 8] backwards 
PASS Range 4 [paras[0].firstChild, 2, paras[0].firstChild, 9] forwards 
PASS Range 4 [paras[0].firstChild, 2, paras[0].firstChild, 9] backwards 
PASS Range 5 [paras[1].firstChild, 0, paras[1].firstChild, 0] forwards 
PASS Range 5 [paras[1].firstChild, 0, paras[1].firstChild, 0] backwards 
PASS Range 6 [paras[1].firstChild, 0, paras[1].firstChild, 1] forwards 
PASS Range 6 [paras[1].firstChild, 0, paras[1].firstChild, 1] backwards 
PASS Range 7 [paras[1].firstChild, 2, paras[1].firstChild, 8] forwards 
PASS Range 7 [paras[1].firstChild, 2, paras[1].firstChild, 8] backwards 
PASS Range 8 [paras[1].firstChild, 2, paras[1].firstChild, 9] forwards 
PASS Range 8 [paras[1].firstChild, 2, paras[1].firstChild, 9] backwards 
PASS Range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] forwards 
FAIL Range 9 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 10 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] forwards 
FAIL Range 10 [detachedPara1.firstChild, 0, detachedPara1.firstChild, 1] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 11 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] forwards 
FAIL Range 11 [detachedPara1.firstChild, 2, detachedPara1.firstChild, 8] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] forwards 
FAIL Range 12 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 13 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] forwards 
FAIL Range 13 [foreignPara1.firstChild, 0, foreignPara1.firstChild, 1] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 14 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] forwards 
FAIL Range 14 [foreignPara1.firstChild, 2, foreignPara1.firstChild, 8] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 15 [document.documentElement, 0, document.documentElement, 1] forwards 
PASS Range 15 [document.documentElement, 0, document.documentElement, 1] backwards 
PASS Range 16 [document.documentElement, 0, document.documentElement, 2] forwards 
PASS Range 16 [document.documentElement, 0, document.documentElement, 2] backwards 
PASS Range 17 [document.documentElement, 1, document.documentElement, 2] forwards 
PASS Range 17 [document.documentElement, 1, document.documentElement, 2] backwards 
PASS Range 18 [document.head, 1, document.head, 1] forwards 
PASS Range 18 [document.head, 1, document.head, 1] backwards 
PASS Range 19 [document.body, 0, document.body, 1] forwards 
PASS Range 19 [document.body, 0, document.body, 1] backwards 
PASS Range 20 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] forwards 
FAIL Range 20 [foreignDoc.documentElement, 0, foreignDoc.documentElement, 1] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 21 [foreignDoc.head, 1, foreignDoc.head, 1] forwards 
FAIL Range 21 [foreignDoc.head, 1, foreignDoc.head, 1] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 22 [foreignDoc.body, 0, foreignDoc.body, 0] forwards 
FAIL Range 22 [foreignDoc.body, 0, foreignDoc.body, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 23 [paras[0], 0, paras[0], 0] forwards 
PASS Range 23 [paras[0], 0, paras[0], 0] backwards 
PASS Range 24 [paras[0], 0, paras[0], 1] forwards 
PASS Range 24 [paras[0], 0, paras[0], 1] backwards 
PASS Range 25 [detachedPara1, 0, detachedPara1, 0] forwards 
FAIL Range 25 [detachedPara1, 0, detachedPara1, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 26 [detachedPara1, 0, detachedPara1, 1] forwards 
FAIL Range 26 [detachedPara1, 0, detachedPara1, 1] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 27 [paras[0].firstChild, 0, paras[1].firstChild, 0] forwards 
PASS Range 27 [paras[0].firstChild, 0, paras[1].firstChild, 0] backwards 
PASS Range 28 [paras[0].firstChild, 0, paras[1].firstChild, 8] forwards 
PASS Range 28 [paras[0].firstChild, 0, paras[1].firstChild, 8] backwards 
PASS Range 29 [paras[0].firstChild, 3, paras[3], 1] forwards 
PASS Range 29 [paras[0].firstChild, 3, paras[3], 1] backwards 
PASS Range 30 [paras[0], 0, paras[0].firstChild, 7] forwards 
PASS Range 30 [paras[0], 0, paras[0].firstChild, 7] backwards 
PASS Range 31 [testDiv, 2, paras[4], 1] forwards 
PASS Range 31 [testDiv, 2, paras[4], 1] backwards 
PASS Range 32 [testDiv, 1, paras[2].firstChild, 5] forwards 
PASS Range 32 [testDiv, 1, paras[2].firstChild, 5] backwards 
PASS Range 33 [document.documentElement, 1, document.body, 0] forwards 
PASS Range 33 [document.documentElement, 1, document.body, 0] backwards 
PASS Range 34 [foreignDoc.documentElement, 1, foreignDoc.body, 0] forwards 
FAIL Range 34 [foreignDoc.documentElement, 1, foreignDoc.body, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 35 [document, 0, document, 1] forwards 
PASS Range 35 [document, 0, document, 1] backwards 
PASS Range 36 [document, 0, document, 2] forwards 
PASS Range 36 [document, 0, document, 2] backwards 
PASS Range 37 [document, 1, document, 2] forwards 
PASS Range 37 [document, 1, document, 2] backwards 
PASS Range 38 [testDiv, 0, comment, 5] forwards 
PASS Range 38 [testDiv, 0, comment, 5] backwards 
PASS Range 39 [paras[2].firstChild, 4, comment, 2] forwards 
PASS Range 39 [paras[2].firstChild, 4, comment, 2] backwards 
PASS Range 40 [paras[3], 1, comment, 8] forwards 
PASS Range 40 [paras[3], 1, comment, 8] backwards 
PASS Range 41 [foreignDoc, 0, foreignDoc, 0] forwards 
FAIL Range 41 [foreignDoc, 0, foreignDoc, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 42 [foreignDoc, 1, foreignComment, 2] forwards 
FAIL Range 42 [foreignDoc, 1, foreignComment, 2] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 43 [foreignDoc.body, 0, foreignTextNode, 36] forwards 
FAIL Range 43 [foreignDoc.body, 0, foreignTextNode, 36] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 44 [xmlDoc, 0, xmlDoc, 0] forwards 
FAIL Range 44 [xmlDoc, 0, xmlDoc, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 45 [xmlDoc, 1, xmlComment, 0] forwards 
FAIL Range 45 [xmlDoc, 1, xmlComment, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 46 [detachedTextNode, 0, detachedTextNode, 8] forwards 
FAIL Range 46 [detachedTextNode, 0, detachedTextNode, 8] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 47 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] forwards 
FAIL Range 47 [detachedForeignTextNode, 7, detachedForeignTextNode, 7] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 48 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] forwards 
FAIL Range 48 [detachedForeignTextNode, 0, detachedForeignTextNode, 8] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 49 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] forwards 
FAIL Range 49 [detachedXmlTextNode, 7, detachedXmlTextNode, 7] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 50 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] forwards 
FAIL Range 50 [detachedXmlTextNode, 0, detachedXmlTextNode, 8] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 51 [detachedComment, 3, detachedComment, 4] forwards 
FAIL Range 51 [detachedComment, 3, detachedComment, 4] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 52 [detachedComment, 5, detachedComment, 5] forwards 
FAIL Range 52 [detachedComment, 5, detachedComment, 5] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 53 [detachedForeignComment, 0, detachedForeignComment, 1] forwards 
FAIL Range 53 [detachedForeignComment, 0, detachedForeignComment, 1] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 54 [detachedForeignComment, 4, detachedForeignComment, 4] forwards 
FAIL Range 54 [detachedForeignComment, 4, detachedForeignComment, 4] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 55 [detachedXmlComment, 2, detachedXmlComment, 6] forwards 
FAIL Range 55 [detachedXmlComment, 2, detachedXmlComment, 6] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 56 [docfrag, 0, docfrag, 0] forwards 
FAIL Range 56 [docfrag, 0, docfrag, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 57 [foreignDocfrag, 0, foreignDocfrag, 0] forwards 
FAIL Range 57 [foreignDocfrag, 0, foreignDocfrag, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
PASS Range 58 [xmlDocfrag, 0, xmlDocfrag, 0] forwards 
FAIL Range 58 [xmlDocfrag, 0, xmlDocfrag, 0] backwards Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges.
Harness: the test ran to completion.
